day 7, part 2
I had to rewrite a lot of how my IntCode processor receives input and
spits out output. It's better now, but this was super obnoxious, and I
had a really hard time understanding that each amp had to pause
execution to wait for input and also pass along output to the next amp
as soon as it had some.